==> notebooks/00_dataset_openmap.ipynb <==
# Core data manipulation and analysis
from pathlib import Path
import pickle
from time import sleep
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')

# Geospatial libraries
import geopandas as gpd
from shapely.geometry import Point, Polygon

import overpy  # For OpenStreetMap data
from urllib.parse import urlencode
import os
os.chdir(r'C:\code\bedford-ubid')

from pymodule.file_ops import read_pickle, write_pickle



bounds = [-73.74815136, 41.16565413, -73.59163084, 41.28549615]


api = overpy.Overpass()

# Construct Overpass query for buildings
query = f"""
[out:json][timeout:60];
(
    way["building"]({bounds[1]},{bounds[0]},{bounds[3]},{bounds[2]});
    relation["building"]({bounds[1]},{bounds[0]},{bounds[3]},{bounds[2]});
);
out geom;
"""
result = api.query(query)


# ids to be loaded
ids_to_process = [x.id for x in result.ways]

ids_to_process

total_id_length = len(ids_to_process)

def try_get_record(way: overpy.Way):
    if len(way.get_nodes(resolve_missing=True)) > 2:  # Valid polygon
        coords = [(float(node.lon), float(node.lat)) for node in way.get_nodes(resolve_missing=True)]
        if coords[0] != coords[-1]:  # Close polygon if needed
            coords.append(coords[0])

        if len(coords) >= 4:  # Valid polygon needs at least 4 points
            polygon = Polygon(coords)
            return {
                'coordinates': coords,
                'geometry': polygon,
                'osm_id': way.id,
                'building_type': way.tags.get('building', 'yes'),
                'source': 'osm',
                # 'way': way
            }
        
    return None

from overpy.exception import OverpassTooManyRequests, OverpassGatewayTimeout
cached_results_file = './datasets/openmap_cache/openmap_cache.pickle'

if not Path(cached_results_file).exists():
    print(f'creating... {cached_results_file}')
    # write_pickle([], cached_id_list_file)
    write_pickle([], cached_results_file)

# cached_id_list = read_pickle(cached_id_list_file)
cached_results = read_pickle(cached_results_file)
print(f'current length: {len(pd.DataFrame(cached_results))}')

cached_ids_list = [x['osm_id'] for x in cached_results]

for w in result.ways:
    id = w.id
    try:
        if id not in cached_ids_list:
            record = try_get_record(w)
            cached_results.append(record)
            write_pickle(cached_results, cached_results_file)

    except KeyError as e:
        print(e)
        sleep(2)
        pass

    except OverpassTooManyRequests:
        print('too many requests, waiting')
        print(f'current length: {len(cached_results)}')
        sleep(5)
        pass
    except OverpassGatewayTimeout:
        print('gateway timeout, waiting')
        print(f'current length: {len(cached_results)}')
        sleep(10)
        pass

cached_results = read_pickle(cached_results_file)
len(cached_results) / len(result.ways)

cached_results[0]


# def try_load(blist, result):
#     found_ids = [x['osm_id'] for x in blist]
#     for way in result.ways:
#         if way.id not in found_ids:
#             sleep(0.1)
#             if len(way.get_nodes(resolve_missing=True)) > 2:  # Valid polygon
#                 coords = [(float(node.lon), float(node.lat)) for node in way.get_nodes(resolve_missing=True)]
#                 if coords[0] != coords[-1]:  # Close polygon if needed
#                     coords.append(coords[0])

#                 if len(coords) >= 4:  # Valid polygon needs at least 4 points
#                     polygon = Polygon(coords)
#                     blist.append({
#                         'coordinates': coords,
#                         'geometry': polygon,
#                         'osm_id': way.id,
#                         'building_type': way.tags.get('building', 'yes'),
#                         'source': 'osm'
#                     })

def try_load(blist, result):
    found_ids = [x['osm_id'] for x in blist]
    for way in result.ways:
        if way.id not in found_ids:
            sleep(0.1)
            if len(way.get_nodes(resolve_missing=True)) > 2:  # Valid polygon
                coords = [(float(node.lon), float(node.lat)) for node in way.get_nodes(resolve_missing=True)]
                if coords[0] != coords[-1]:  # Close polygon if needed
                    coords.append(coords[0])

                if len(coords) >= 4:  # Valid polygon needs at least 4 points
                    polygon = Polygon(coords)
                    blist.append({
                        'coordinates': coords,
                        'geometry': polygon,
                        'osm_id': way.id,
                        'building_type': way.tags.get('building', 'yes'),
                        'source': 'osm'
                    })

                    

# try_bldg = []
# try_load(try_bldg, result)



buildings = []
current_length = 0
missing_ids_length = total_length - current_length


while missing_ids_length != 0:
    current_length = len(buildings)
    missing_ids_length = total_length - current_length
    print('trying...')
    print(f'{missing_ids_length} missing')
    try:
        try_load(buildings, result)
    except:
        pass

buildings

buildings


# found_ids = [x['osm_id'] for x in buildings]

# len_pre = len(buildings)
# for way in result.ways:

#     if way.id not in found_ids:
#         print(f'trying to load: {way.id}')
#         sleep(1)
#         if len(way.get_nodes(resolve_missing=True)) > 2:  # Valid polygon
#             coords = [(float(node.lon), float(node.lat)) for node in way.get_nodes(resolve_missing=True)]
#             if coords[0] != coords[-1]:  # Close polygon if needed
#                 coords.append(coords[0])
            
#             if len(coords) >= 4:  # Valid polygon needs at least 4 points
#                 polygon = Polygon(coords)
#                 buildings.append({
#                     'geometry': polygon,
#                     'osm_id': way.id,
#                     'building_type': way.tags.get('building', 'yes'),
#                     'source': 'osm'
#                 })
    

# len_post = len(buildings)
# print(f'{len_post - len_pre} added. total length: {len_post}')
            


buildings_gdf = gpd.GeoDataFrame(buildings, crs='EPSG:4326')


# Function to get OpenStreetMap buildings

    print("Querying OpenStreetMap for buildings...")
    
    
    
    # Convert to GeoDataFrame
    
            

print("Building data access functions defined successfully!")
# Convert to WGS84 for web mapping if needed


## 4. Data Exploration and Preprocessing

## 5. Spatial Join Operations

Perform spatial joins to determine which buildings fall within each parcel boundary.
